Rediseño del middleware para utilizar dependencias no deprecated
